+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.2
+
+* new MetricAnomaly that allows swarming to optimize anomaly parameters (to match expected anomaly percentage on data).
+* allow named models (in description.py: config{ 'name': xxx,}) and access them globally from GlobalDict.get(name).
+
 ## 0.2.1
 
 * Moved data pkg_resource data into nupic/datafiles.